make pass InputTypeTest

+++ src/jebl/evolution/characters/DiscreteCharacter.java
@@ -96,7 +96,7 @@ public class DiscreteCharacter extends BEASTObject implements Character{
 	 * 
 	 * @return the number of possible states for the character
 	 */
-	public double getNumOfStates(){ return numOfStates; }
+	public Integer getNumOfStates(){ return numOfStates; }
 	
 	/**
 	 *
